Created multer middleware and cloudinary util and fixed bugs in register controller.

+++ Campus_Engage/Backend/src/utils/cloudinary.js
@@ -8,15 +8,23 @@ v2.config({
 });
 
 const uploadOnCloudinary = async (localFilePath) => {
+    console.log('[CLOUD] start upload:', localFilePath, new Date().toISOString());
 
     try {
         if (!localFilePath) return null;
         const response = await v2.uploader.upload(localFilePath, { resource_type: "auto" });
-        fs.unlinkSync(localFilePath);
+         console.log('[CLOUD] uploaded OK:', response.url);
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+             console.log('[CLOUD] deleted file:', localFilePath);
+        }
         console.log("File has been uploaded: ", response.url);
         return response;
     } catch (error) {
-        fs.unlinkSync(localFilePath);
+        console.error('[CLOUD] upload error:', err);
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
         console.log(error);
         return null;
     }
